requirements.txt
openai>=0.27.0
pydantic>=2.0
python-dotenv>=1.0
pytest>=7.0
